requirements.txt
allure-pytest==2.13.5
pytest==8.3.4
selenium==4.29.0
webdriver-manager==4.0.2

